# Refactor VisitType to be fully iterative