--- a/src/Compilers/Core/Portable/Collections/ArrayBuilderExtensions.cs
+++ b/src/Compilers/Core/Portable/Collections/ArrayBuilderExtensions.cs
@@ -6,6 +6,7 @@
 
 using System;
 using System.Collections.Immutable;
+using System.Diagnostics.CodeAnalysis;
 using Microsoft.CodeAnalysis.PooledObjects;
 
 namespace Microsoft.CodeAnalysis
@@ -160,6 +161,18 @@ namespace Microsoft.CodeAnalysis
             var e = builder.Peek();
             builder.RemoveAt(builder.Count - 1);
             return e;
+        }
+
+        public static bool TryPop<T>(this ArrayBuilder<T> builder, [MaybeNullWhen(false)] out T e)
+        {
+            e = default;
+            if (builder.Count == 0)
+            {
+                return false;
+            }
+
+            e = builder.Pop();
+            return true;
         }
 
         public static T Peek<T>(this ArrayBuilder<T> builder)