Previously, `VisitType` attempted to implement half-tail recursion: for cases where there was a single nested type, such as pointer types and array types, we would use an iterative pattern. However, when visiting a type's containing type or when there were nested types that had potentially multiple children, such generic type arguments or function pointer types, we would fall back to standard recursion. These latter cases, particularly generics, can be deeply nested and often present issues in our smoke tests. I've rewritten `VisitType` to be fully iterative in all cases, using a manual stack instead of recursion to handle deeply nested type cases.