@333fred 2020-06-19
Perf Benchmarks. @agocke to confirm I'm reading this data right, but it appears that manual iteration here is very marginally slower than the recursive method (something like between 10-50 ms slower on `CompileMethodsAndEmit`, and 10-20 ms slower on `SerializeMetadata`), while not affecting memory allocations in any significant fashion (a few thousand KB at most).

Before:

|                      Method |        Job | UnrollFactor |        Mean |    Error |   StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 |  Allocated |
|---------------------------- |----------- |------------- |------------:|---------:|---------:|------------:|------------:|------------:|------------:|-----------:|----------:|-----------:|
|       CompileMethodsAndEmit | Job-MWBLNZ |           16 |  4,872.5 ms | 31.45 ms | 27.88 ms |  4,865.3 ms |  4,832.1 ms |  4,924.6 ms |  90000.0000 | 26000.0000 |         - |  541.54 MB |
|           SerializeMetadata | Job-MWBLNZ |           16 |    274.1 ms |  2.73 ms |  2.81 ms |    274.5 ms |    268.5 ms |    278.8 ms |   3000.0000 |  1000.0000 |         - |   33.56 MB |
|              GetDiagnostics | Job-JXYEFR |            1 |  3,737.7 ms | 26.25 ms | 24.55 ms |  3,732.9 ms |  3,710.1 ms |  3,779.6 ms |  81000.0000 | 20000.0000 |         - |  423.77 MB |
| GetDiagnosticsWithAnalyzers | Job-JXYEFR |            1 | 12,476.5 ms | 63.07 ms | 52.67 ms | 12,491.8 ms | 12,389.0 ms | 12,565.7 ms | 218000.0000 | 61000.0000 | 3000.0000 | 1312.05 MB |

After:

|                      Method |        Job | UnrollFactor |        Mean |    Error |   StdDev |      Median |         Min |         Max |       Gen 0 |      Gen 1 |     Gen 2 |  Allocated |
|---------------------------- |----------- |------------- |------------:|---------:|---------:|------------:|------------:|------------:|------------:|-----------:|----------:|-----------:|
|       CompileMethodsAndEmit | Job-UTEKXA |           16 |  4,933.3 ms | 22.45 ms | 19.90 ms |  4,931.6 ms |  4,906.0 ms |  4,978.2 ms |  90000.0000 | 24000.0000 |         - |  542.55 MB |
|           SerializeMetadata | Job-UTEKXA |           16 |    296.6 ms |  2.78 ms |  2.60 ms |    298.0 ms |    292.7 ms |    300.9 ms |   3000.0000 |  1000.0000 |         - |   33.56 MB |
|              GetDiagnostics | Job-MJXNWV |            1 |  3,747.5 ms |  7.47 ms |  5.83 ms |  3,748.5 ms |  3,735.9 ms |  3,754.0 ms |  82000.0000 | 21000.0000 |         - |  423.87 MB |
| GetDiagnosticsWithAnalyzers | Job-MJXNWV |            1 | 12,507.6 ms | 74.13 ms | 69.34 ms | 12,480.5 ms | 12,421.4 ms | 12,616.4 ms | 216000.0000 | 60000.0000 | 2000.0000 | 1301.28 MB |