@333fred 2020-06-19
Closing this out. We're not going to be able to take this approach. I'll instead go for a modest refactoring that will improve the tail recursive nature of the existing `VisitTypes` scenario.